refactoring

+++ loms/internal/model/stock.go
@@ -1,6 +1,6 @@
 package model
 
 type Stock struct {
-	Available uint64
-	Reserved  uint64
+	Available uint16
+	Reserved  uint16
 }
