+++ loms/internal/service/stock/stock.go
@@ -15,7 +15,7 @@ func New(rep repository.Stock) *stock {
 	return &stock{rep: rep}
 }
 
-func (s *stock) GetAvailable(ctx context.Context, sku model.SKU) (uint64, error) {
+func (s *stock) GetAvailable(ctx context.Context, sku model.SKU) (uint16, error) {
 	st, err := s.rep.FindBySKU(ctx, sku)
 	if err != nil {
 		return 0, err
@@ -23,14 +23,14 @@ func (s *stock) GetAvailable(ctx context.Context, sku model.SKU) (uint64, error)
 	return st.Available, nil
 }
 
-func (s *stock) AddReserve(ctx context.Context, sku model.SKU, count uint64) error {
+func (s *stock) AddReserve(ctx context.Context, sku model.SKU, count uint16) error {
 	return s.rep.AddReserve(ctx, sku, count)
 }
 
-func (s *stock) CancelReserve(ctx context.Context, sku model.SKU, count uint64) error {
+func (s *stock) CancelReserve(ctx context.Context, sku model.SKU, count uint16) error {
 	return s.rep.CancelReserve(ctx, sku, count)
 }
 
-func (s *stock) ShipReserve(ctx context.Context, sku model.SKU, count uint64) error {
+func (s *stock) ShipReserve(ctx context.Context, sku model.SKU, count uint16) error {
 	return s.rep.ShipReserve(ctx, sku, count)
 }
